# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Messenger Gateway API
  version: 1.0.0
  description: REST API for mobile client (MVP)
servers:
  - url: http://localhost:8080
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /v1/auth/ws-ticket:
    post:
      tags: [Auth]
      summary: Issue one-time WS ticket
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Ticket issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  ticket:
                    type: string
                  ttl:
                    type: integer
                    example: 60
        "401":
          description: Unauthorized
...
  /ws:
    get:
      tags: [WS]
      summary: WebSocket endpoint
      description: Connect using one-time ticket `ws://host/ws?ticket=...`.
      parameters:
        - name: ticket
          in: query
          required: true
          schema:
            type: string
      responses:
        "101":
          description: WebSocket upgrade
        "401":
          description: Missing, invalid, or expired ticket